# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Trade AI Gateway API
  version: "0.3.0"
  description: >-
    Contract stubs for Phase-1 MVP; BE may return 501 until implement;
    x-mvp true = In-MVP; FE must not invent fields.
servers:
  - url: http://localhost:8000
    description: Local gateway
security:
  - bearerAuth: []
...
paths:
...
  /v1/accounts:
    post:
      operationId: postAccounts
      summary: Create a trading account
      tags: [Accounts]
      x-mvp: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreate"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          $ref: "#/components/responses/NotImplemented"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/accounts/{account_id}/api-keys:
    parameters:
      - $ref: "#/components/parameters/AccountId"
    post:
      operationId: postAccountApiKeys
      summary: Register exchange API credentials
      tags: [Accounts]
      x-mvp: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyCreate"
      responses:
        "201":
          description: API credentials registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyMasked"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          $ref: "#/components/responses/NotImplemented"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/strategies:
    get:
      operationId: getStrategies
      summary: List strategies
      tags: [Strategies]
      x-mvp: true
      parameters:
        - $ref: "#/components/parameters/AccountIdQuery"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/StrategyStatus"
      responses:
        "200":
          description: Strategies returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          $ref: "#/components/responses/NotImplemented"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: postStrategies
      summary: Create a strategy
      tags: [Strategies]
      x-mvp: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyCreate"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          $ref: "#/components/responses/NotImplemented"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/strategies/{strategy_id}:
    parameters:
      - $ref: "#/components/parameters/StrategyId"
    patch:
      operationId: patchStrategy
      summary: Update a strategy
      tags: [Strategies]
      x-mvp: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyPatch"
      responses:
        "200":
          description: Strategy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          $ref: "#/components/responses/NotImplemented"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AccountId:
      name: account_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AccountIdQuery:
      name: account_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
    StrategyId:
      name: strategy_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The caller is not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotImplemented:
      description: Contract stub is not implemented by the backend
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ErrorResponse:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
        message:
          type: string
        trace_id:
          type: string
          format: uuid
        details:
          type: array
          items:
            type: object
            required: [field, reason]
            properties:
              field:
                type: string
              reason:
                type: string
...
    AccountCreate:
      type: object
      required: [name, exchange, market_type]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        exchange:
          type: string
          minLength: 1
        market_type:
          type: string
          enum: [spot, futures]
        testnet:
          type: boolean
          default: true
    Account:
      type: object
      required:
        [id, name, exchange, market_type, testnet, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        exchange:
          type: string
        market_type:
          type: string
          enum: [spot, futures]
        testnet:
          type: boolean
        status:
          type: string
          enum: [active, disabled, error]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ApiKeyCreate:
      type: object
      required: [label, api_key, api_secret]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 100
        api_key:
          type: string
          writeOnly: true
          minLength: 1
        api_secret:
          type: string
          writeOnly: true
          minLength: 1
        passphrase:
          type: string
          writeOnly: true
          minLength: 1
    ApiKeyMasked:
      type: object
      required: [id, account_id, label, masked_api_key, created_at]
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        label:
          type: string
        masked_api_key:
          type: string
          example: "****MASKED"
        created_at:
          type: string
          format: date-time
        last_validated_at:
          type: string
          format: date-time
          nullable: true
    StrategyStatus:
      type: string
      enum: [draft, active, paused, stopped]
    StrategyCreate:
      type: object
      required: [account_id, name, symbol, timeframe]
      properties:
        account_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 120
        symbol:
          type: string
          minLength: 1
        timeframe:
          type: string
          enum: [1m, 5m, 15m, 1h, 4h, 1d]
        status:
          $ref: "#/components/schemas/StrategyStatus"
        max_position_size:
          type: number
          format: double
          minimum: 0
        stop_loss_percent:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
          maximum: 100
    StrategyPatch:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        status:
          $ref: "#/components/schemas/StrategyStatus"
        timeframe:
          type: string
          enum: [1m, 5m, 15m, 1h, 4h, 1d]
        max_position_size:
          type: number
          format: double
          minimum: 0
        stop_loss_percent:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
          maximum: 100
    Strategy:
      type: object
      required:
        [id, account_id, name, symbol, timeframe, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        name:
          type: string
        symbol:
          type: string
        timeframe:
          type: string
        status:
          $ref: "#/components/schemas/StrategyStatus"
        max_position_size:
          type: number
          format: double
        stop_loss_percent:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...